create shortcut for initials
setup shortcuts for ALT + Numpad{num} to send input of the value of the initials from the current gui state followed by " P-0{num}". Then also followed by an input of shift+enter and enter, so in excel the data is entered without going to the next cell, as shift+enter goes 1 cell up, and enter goes 1 cell back down.