add button to open main "printers and faxes" window
opens the PrinterAndFaxes window using Shell and the CLSID for the window. Starts minimized and will wait for the window to be found. If not found, it will show an error message and return from the function to prevent opening other windows that are dependent on the PrinterAndFaxes window to be open. Setup detecting of hidden windows in case it is open but hidden, and set title match to only match the exact title, which is "Printers - File Explorer".